backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.11.0
statsmodels>=0.14.0
pingouin>=0.5.3
pyreadstat>=1.2.4
openpyxl>=3.1.2
polars>=0.20.2
jinja2>=3.1.2
weasyprint>=60.2
python-docx>=1.1.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
ruff>=0.1.6
